db: create indexes on fields we are going to use in filters
Indexes will be created in the "background": read / write operations
are allowed on the documents of the collection, while the new indexes
are being built.

At the same time, one can't drop a collection, while an index build
is in progress, thus, replace all drop()'s in tests with remove()'s,
so that we don't drop the collection between tests and don't have
to rebuild indexes again.