Pass concurrency=greenlet to coverage.py
So far coverage wrongly reports missed statements, and shows a bunch of
create_index calls as unevaluated. We know for sure they have been
executed, since otherwise tests will fail.

According to coverage documentation, it easily can be confused if greenlet
is used. In that case we should explicitly tell coverage that
concurrency model is greenlet.

Unfortunately, this is the case. Motor uses greenlet under the hood, and
passing concurrency=greenlet to coverage solves the issue.