Add marker based pagination support to /snippets
It's now possible to pass additional ``limit`` and ``marker`` query
parameters to /snippets endpoint, e.g.:

    GET /snippets?limit=10&marker=1234

where ``limit`` will be an upper bound for the number of snippets to
be returned and ``marker`` is the ``id`` of the last snippet seen on
the previous page. Snippets are ordered by the creation order (newer
ones will be returned first).

Marker based pagination is used instead of limit/offset one, as the
performance of the latter degrades quickly as the number of entries
grows.