Implement API versioning via X-Api-Version header
It's naive to believe that API doesn't evolve. We live in brutal real
world where everything tends to be changed, and XSnippet API is not an
exception here. There are different approaches to version API, but we
chose to use "X-Api-Version" header. If nothing is passed, the latest
version is used.