Untie TestResource from XSnippet application
We shouldn't use XSnippet application created by `create_app` factory
because we can't rely on its router. Apparently, `add_route` method is
not a part of router's interface [1]. In light we want to implement
API versioning through custom router, it's not good to rely on methods
out of router interface.

[1]: https://github.com/KeepSafe/aiohttp/blob/v0.22.4/aiohttp/abc.py#L10-L15