POST /snippets - validate input data
In order to prevent saving incorrect snippet in database we've got to
reject any requests with passed malformed snippet. We use Cerberus
library to check input data against Snippet schema. The 'snippet.syntax'
property, on the other hande, has a special treating. It depends on
configuration option ('snippet.syntaxes') and if the last one is
specified, only those values are allowed.