Respond with 'Vary' HTTP header
The Vary HTTP response header determines how to match future request
headers to decide whether a cached response can be used rather than
requesting a fresh one from the origin server. It is used by the server
to indicate which headers it used when selecting a representation of a
resource in a content negotiation algorithm.

This commit adds three Varying headers: Accept, Accept-Encoding and
X-Api-Version.