Add authentication middleware
The proposed middleware performs authentication of user requests by
validating tokens passed in X-Auth-Token header. The values are
expected to be in JSON Web Token format (RFC 7519).

The parsed token value (if validation succeeded) is then stored in
the request object to be used later in resource handlers.

This is the first patch in the series of authentication related
changes. Subsequent patches will introcude the token format
and corresponding service layer functions.