Replace 'X-Auth-Token' with 'Authorization' header
For a long time non-standard header fields were conventionally marked by
prefixing the field name with 'X-', but this convention was deprecated
in June 2012 [1] because of the inconveniences it caused when non-standard
fields became standard.

So the obvious move is to strip 'X-' prefix and go with 'Auth-Token' header
instead. However, it's still non-standard header. Fortunately, there's
an 'Authorization' header that's commonly used for passing tokens and is
a recommended way to pass a JWT [2] token.

This commit changes tokens passing from

    X-Auth-Token: <token>

to

    Authorization: bearer <token>

[1] https://tools.ietf.org/html/rfc6648
[2] https://jwt.io/introduction/