Use Cerberus to validate GET parameters
Instead of manual casting and checking each GET parameter for a valid
value, it's much readable and extendible to use Cerberus validation
library.

So far we have only two GET parameters which are used for pagination:

  - `limit`
  - `marker`

They are both limited to be a positive integers with some min/max
constraints. So the Cerberus schema is introduced for them in order
to check those constraints.