db: create db connection and indexes at application startup
We currently mix synchronous and asynchronous code in
create_connection(), that leads to sporadic warnings in tests due to
the fact that `on_done` future callbacks are called after the event
loop instance is already closed. Apparently these futures returned
by create_index() are not awaited by the event loop as it is, so we'd
better do it explicitly, rather than hope that callbacks will be
called in time before the event loop is closed.

Delay creation of a DB connection till the application startup moment
(i.e. run_app() in aiohttp) for the sake of consistency, so that
all setup steps are performed in one single place. The tests are
modified accordingly.

With this change applied I no longer run into the default limit for
the number of connections opened to mongod at the same time when
running tox -epy36.