Minor enhancements in tox.ini
May God and @malor forgive me, but I made several unrelated changes
to `tox.ini` in one commit. Why? I think changing three lines in
several commits including the very one that will be changed in each
of them, rather a silly idea; so there we are.

The commit accommodates the following changes:

 * `py.test` executable is deprecated in favor of `pytest`, however,
   it was decided to use Python module notation instead as it doesn't
   rely on binary name.

 * Both `pytest` and `flake8` might be available globally on the machine
   (e.g. pytest on Travis CI), so let's use `{envpython}` notation
   instead of binary names in order to avoid ambiguous situations.

 * The `--strict` flag is passed to `pytest` to treat warnings as
   errors.

 * Both `pytest` and `flake8` run for the whole project instead of
   just xsnippet_api/ and tests/ folders; this allows us to have some
   tests inside contrib/ directory (if any).