Add `GET /syntaxes` endpoint
This new endpoint allows to retrieve the list of supported snippet
syntaxes, that is defined in the config file. We default to an
empty list, though, which means the API will accept any syntax value
on creation of a new snippet.

Update the config we use in ansible deployment to default to list
of syntaxes supported by CodeMirror.

Put the test fixtures into conftest.py, so that they can be reused
in tests of different API resources.