Remove router assignment via app._router property
We use to set VersionRouter instance on application later in the code by
exploiting private attribute. The need of this hack comes from the time
when UrlDispatcher (router) was needed to be aware of the application,
and the application was needed to be aware of router on creation time.

In aiohttp v2.3.5 (we depend on) it's no more the case. So let's remove
the hack and use intended design.