Improve the efficiency of pagination
The idea of marker-based pagination is that users pass the id of
the last item on a page that was returned in order to obtain the
next page.

On the db side we use the given id to find the "specimen" document
first and then use the "key" value (in our case we sort by value of
`created_at` attribute) to efficiently find the position of the
specimen document in a db index and then select the following `$limit`
documents (note, that we use btree indexes, which store all the keys
in some order, so the db just needs to follow btree node pointers -
no sorting is actually performed).

In order to avoid the ambiguity of multiple documents with the same
value of `created_at` attribute, we used a compoud sorting key -
`(_id, created_at)`. The problem with that was that we did not have
a compound index on those attributes, so Mongo query plan for a list
of snippets on some page looked like:

    1) scan the _id index to find keys where _id < $specimen_id
    2) fetch all the documents with _id's from 1) and also filter by
       created_at <= $specimen_created_at
    3) sort the fetched documents by (_id, created_at) in descending order
    4) take the first $limit documents and return them

^ the problem is that on the first pages (which are the viewed the most)
many documents would satisfy the conditions in 1) and 2), which is
exactly what we tried to avoid.

By having a compound index on `(created_at, _id)` and changing the
sorting key we get the following query plan:

    1) find the position of the specimen document in the compound index
    2) fetch the first $limit documents that that have (created_at, _id)
       less than (specimen_created_at, specimen_id)

I gave it a try on a test collection with 10k snippets and retrieved
the 20-item page that goes after the snippet with id equal to 9800
(i.e. 9800 are "older" than this one):

    1) w/o this change: 25ms, 9800 keys examined
    2) with this change: 1ms, 20 keys examined

Note, due to the nature of btree indexes we still use the very same
compound index for prefix queries by `created_at` only when `_id`
is not specified.