Use explicitly set API version when not passed
We used to fallback to latest stable API version when 'Api-Version'
header was omitted. While it made sense when we were using version
discovery mechanism, it makes a little sense now when all versions
are statically defined.

Since now it's more garbage and complication, not saying it doesn't
give us control on which version to serve by default, let's remove
this function now.