Use DI to inject database/conf where needed
So far database connection available as resource's 'db' attribute and we
passed it down to service layer.  Similarly, we extract 'conf' out from
application which is extracted out of request. It works, and it works
good, however code readability is not so good as we would like.

In order to improve code readability and provide some sort of connection
points and ask for a 'database' and 'conf' only where needed, let's try
to use dependency injection framework to provide those.

Unfortunately, some barely related changes are included in this PR, so
may Roman forgive me this. :)

Change-Id: If97fb3cc6ea7ae173384682aab9e3cf1a1c4dc86