Remove /app directory (sources) from Docker image
Since we install the app in the container by running `pip install` the
app will be copied over to some `/usr/lib/python3/site-packages` place.
That basically means we don't the sources in the container afterwards,
so why just not remove them to avoid ambiguousness about which code
is actually used to run the app?