Generate secret in main rather than in middleware
There are couple of reasons of doing so but the main one is to make
middleware to be responsible for validating tokens only, moving
responsibility of getting secret (reading from config or generating one)
from middleware (as it's supposed to be as general as possible) to
consumers.

It'd make even more sense once we implement /auth endpoint that would be
responsible for issuing JWT tokens, meaning that the middleware is not
the only owner of authentication.