Do not default 'syntax' to 'text' if not passed
We used to have sort of weird behaviour by defaulting snippet's 'syntax'
to 'text' if it wasn't specified at the first place. While it'd make
sense on certain situations, it's something we want to avoid in
XSnippet, because we might want to know whether it's something submitted
by user or not.

Clearly there are two options to go with:

 (a) Require 'syntax' to be explicitly passed in POST and PUT requests.
 (b) Assume 'syntax' is 'None' if not passed.

In case of option (a) we have an open question: what to do if we didn't
set a list of allowed syntaxes? Should we force users to pass any syntax
even if it doesn't matter?

In case of option (b) an open question is opposite: what to do if we set
a list of allowed syntaxes? Should we ignore this implicit 'syntax'
value and allow to store 'None'?

Despite not having a good way to go, we must come up with some solution.
So this commit implements option (b).