docker: use python:slim as a base to reduce the image size
python:latest size is ~900 Mb, which sounds a lot for such a simple
app as xsnippet-api; we don't really need all that stuff. Use
python:slim image as a base instead, that is much smaller (~130 Mb).

As we still need some external tools like git to be pre-installed
in order to produce an xsnippet-api wheel, we use the new Docker
multi-stage builds feature: the first stage is responsible for
collecting all the necessary Python wheels (including xsnippet-api
itself); the second stage only copies the artifacts and installs
them to a slim image.

See Docker docs for details:

https://docs.docker.com/develop/develop-images/multistage-build/#use-multi-stage-builds