Update Motor to 2.0
Motor 2.0 drops support for MongoDB 2.6 and adds supports MongoDB 4.0
features, including multi-document transactions, and change stream
notifications on entire databases or entire MongoDB servers. It adds
support for Python 3.7. This version of Motor requires PyMongo 3.7 or
later.

Unfortunately, this change also drops support of PyPy3 because the
latter is compatible with Python 3.5 only.