Parametrize a set of supported syntaxes on start
Dealing with snippets, XSnippet API has a notion of syntaxes, one may
consider using to highlight the snippet on the web or perform some
actions in other clients. These syntaxes may differ based on supported
syntaxes in XSnippet Web, or needs of others when this service is ran on
premises.

This patch adds parametrization via Rocket's built-in facilities, such
as Rocket.toml or ROCKET_SYNTAXES environment variable. The variable
must be a list of strings.

  $ ROCKET_SYNTAXES=["rust", "python"] cargo run

By default, no syntaxes are used.