Use alembic for schema migrations
Pros:

* we do not have to write raw SQL; alembic allows us to target more
than one database backend using the same code (e.g. so that we can
run tests on SQLite locally)

* the directory with migration scripts can be placed anywhere within
the repo, unlike with diesel_cli that has not been released with
this feature yet :)

* alembic is more mature and feature-rich

Cons:

* one more tool in a different programming language (Python)

* one more step to generate Diesel schema macros (diesel_cli combines
this step with running of the migration scripts). We will need to run
the following command every time we change the database schema:

    $ diesel print-schema > src/storage/sql/schema.rs

* we won't have the option of not shipping the migration scripts with
the code

Usage:

$ python -m pip install alembic psycopg2
$ export DATABASE_URL=postgres://user:password@host:port/database
$ export DATABASE_URL=sqlite:///path/to/db
$ alembic upgrade head  # upgrade to the latest revision
$ alembic downgrade base  # downgrade to the base (empty) revision
$ alembic current  # print the current revision