The initial implementation of the storage layer
Trait Storage defines a basic CRUD interface for storing/loading
snippets from a persistent database. The only implementation planned
as of now is SqlStorage, which only supports PostgreSQL at the moment,
while in the future it should be relatively easy to extend it to
support SQLite and MySQL if desired (we would need to stop using
INSERT ... ON CONFLICT DO ... , though).