gh-actions: use psycopg2-binary package
psycopg2 is distributed in sources and hence is required to be compiled
and linked against various libraries. Apparently, there are some
compilation issues on macOS related to not being able to find libssl.
Fortunately, there's an offical package with precompiled binaries
(wheels) so let's use to workaround the issue and even to speed up jobs'
time.