Reimplement pagination to use markers instead of limit+offset
Pagination based on limit+offset does not scale with the number of
items (because offset is O(N)). Rework the code to use markers
instead, which allows us to find the required page in O(log N).

Only basic tests are added. We will test this properly in the
integration tests when we wire this code up to the API.