Make Snippet and Changeset serializable
In order to return Snippets in API responses, we need to make them
serializable first. Rocket can handle serialization to JSON, but it
requires the types to implement `serde::Serializable`.

In theory, we could just auto-derive that trait for our structs, but
in practice it is more complicated than that, as `created_at` and
`updated_at` are of type `DateTime<Utc>`, and the latter does not
implement `serde::Serializable`.

Rust has this fair, but sometimes annoying requirement, that you can't
implement foreign traits for foreign types. So the way to work around
that is to create a wrapper struct around `DateTime<Utc>` from the chrono
crate. This has no runtime/memory overhead, but it is a little verbose
(because we need to handle type conversions).