Make Gabbit tests run DB migrations
As briefly discussed in the comments to #112, we want to make the Gabbit
tests more self-sufficient: given a database URL, they should be able to
start an instance of xsnippet-api, including running the DB migrations
by the means of alembic.

While we are at this, fix the warning emitted by SQLAlchemy: the name of
the DB API driver has been changed from `postgres` to `postgresql`.
Diesel is happy with both, so no change there.