Use chrono with "serde" feature enabled
This allows us not to call .to_rfc3339() explicitly and rely on
chrono choosing the correct serialization format.

One difference is that the timestamps in UTC now use the Z suffix
instead of encoding the HOURS:MINUTES offset (+00:00). Both are
valid representations according to the standard. Account for that
in the regexp that we use in tests.