Add support for JWT-based authentication and authorization
Allow users to send authorized requests by passing a signed JSON Web
Token value. The default auth provider is Auth0, but any provider
that uses RSA for signing and validating tokens can be configured
instead (which is what we do in functional tests).

XSnippet API will configure the token validator based on the values
of the following environment variables:

* ROCKET_JWT_AUDIENCE (e.g. https://api.xsnippet.org)
* ROCKET_JWT_ISSUER (e.g. https://xsnippet.eu.auth0.com)
* ROCKET_JWT_JWKS_URI (e.g. https://xsnippet.eu.auth0.com/.well-known/jwks.json)

The latter specifies the location of JSON Web Key Set, which contains
the _public_ keys required for validating token signatures. These keys
are read once when the application starts; no additional requests to
the auth provider APIs are performed. Only the asymmetric-key crypto
algorihms like RS256 are supported at the moment.

As for the users and permissions, the proposed model is very basic
right now:

* requests w/o the Authorization header are considered to be performed
by user "Guest". This user does not have any privileges besides the
ability to create and retrieve anonymous public snippets

* requests w/ the Authorization header that pass token validation
are performed by user "Authenticated". An authorized user has a unique
identifier (name), and can have some additional permissions configured.
Authenticated users will (when the corresponding functionality is implemented)
be able to create, update, retrieve, and delete privately owned snippets
with the controlled level of visibility

* Auth providers like Auth0 allow to assign additional permissions
which will be included in the signed tokens (e.g. permission "admin"
can be added to for service owners).

In order to not rely on some predefined users and shared credentials,
the test fixture acts as a very simple auth provider that generates
a single-use RSA keypair, and a set of tokens that are passed to the
Gabbi scenarios.