Allow to "import" snippets via POST /v1/snippets/import
Normally when new snippets are created, users are not allowed to
specify the values of some fields which are automatically generated,
e.g. `id`, `created_at`, and `updated_at`. This would, however, be
be very useful for "importing" the state of the previous incarnation
of XSnippet API that is currently deployed to prod (so that we can
leverage the API for this instead of INSERT'ing values directly to
the database).

Importing snippets should not be allowed to anyone, but only to users
with the special `import` permission.