Fix the edge case of pagination being broken for snippets imported from Mongo
When listing snippets, we add filters on `created_at` and `id` to fetch
the next page. The latter is only needed because the value of
`created_at` is not guaranteed to be unique, so the marker can appear in
the next page again. In practice, because we use microsecond precision for
datetime fields, duplicates are only expected in tests and, potentially,
in snippets imported from Mongo that have second precision.

As is, the code did not work correctly in the edge case when snippets
were imported from the old API instance in the reverse (starting from
the latest snippets) order, because the assumption, that both `created_at`
and `id` monotonically increase together, did not hold true: the additional
filter `id < $MARKER_ID` produced an empty result, as "older" (in terms of
the value of `created_at`) entries had larger integer ids.

The fix is to change the WHERE clause to compare the value of
`created_at` first, and only use `id` to resolve the ties, rather
than for every comparison.