Add a GET /snippet/:id that allows to retrieve the raw content of a snippet as plain text
This route can be used to implement the "raw" URIs in the UI w/o the
need to support the corresponding nginx Lua rules.

The way we have written out content negotiation system requires us to
either add support for a new content type to *all* routes (which we
don't want to do at this point), or to use a new request guard to make
Rocket dispatch the request to a correct function (which is a bit
cumbersome but still easier).