Update to Rocket 0.5
We are considering ditching Rocket for axum, but, in the meantime,
let's at the very least update to the latest Rocket version so that:

* we can fix all the security vulnerabilities Dependabot is complaining about

* we can use stable Rust instead of nightly

There are many backward incompatible changes, but most were due to async
and new lifetimes sprinkled everywhere now.

One annoying bit is that we have to temporarily give up on our tracing
integration: request span can no longer be created on the same thread as
Tokio can move futures between threads, and the span must be created on
the top-level "request" future, i.e. in Rocket itself. The code
compiles, but it is disabled.