Migrate to diesel-async
Now that we are using asynchronous web-framework and routes, we want to
get rid of blocking database calls. Diesel is asynchronous, but there is
this diesel-async extenstion that we can use, so let's migrate to that.
This is mostly straightforward as diesel-async provides direct
replacements for Diesel APIs, but we still have to put async everywhere
and play by Future's rules.

The most notable change is a different connection pool implementation,
as diesel-async does not support the one we used before. The API is
almost the same, but default max pool size would probably insufficient,
so we can override that.