Fix asset uploading
`gh release upload` allows to set the display label, but, as it turns
out, that does not change the names of the assets in the release, and
the names are required to be unique. In our case, the names of the
Linux and MacOS binaries clash, and the slower of them will fail to
upload.