Build and link with openssl statically
Dynamic linking with openssl does not work well with cross-compilation:
the Docker image `cross` uses for aarch64 is based on Ubuntu Xenial
that has libssl1.0.0, while the ones available on Raspbian (and other
modern distros) are libssl1.1 and libssl3.

Instead, we can use the feature of `openssl-sys` that allows to build
and statically link openssl to avoid the hassle with shared object
versions that vary between Linux distros.

This makes Cross.toml completely unnecessary.