ci: fix build error on windows
See https://github.com/sfackler/rust-openssl/issues/2149.

Apparently, in the Windows runner, commands executed in a `run` block
will use the version of Perl built into MinGW instead of the one
installed in the system, which breaks the openssl build because one of
the expected Perl modules can't be found.

There are different ways to fix this. The one suggested in this PR is
returning back to using the cargo action for "normal" (not cross
compilation) builds like it used to work before. My understanding is
that it helps by running cargo directly as opposed to running it from
Bash.

Alternatively, we could bypass the corresponding build steps in openssl
by providing locations of openssl source code via environment variables
like it is suggested in the rust-openssl issue linked above.